requirements/common.txt
Django==1.6.5
South==1.0
psycopg2==2.5.3
wsgiref==0.1.2
django-localflavor==1.0
